backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
pydantic==2.5.2
google-generativeai==0.3.1
python-dotenv==1.0.0
pytest==7.4.3
gunicorn==21.2.0
